Moved the item count form to its tab.
Also fixed the width of forms and cleaned up code a little.
