+++ src/features/WorkerRegistrationForm/WorkerRegistrationForm.tsx
@@ -20,7 +20,7 @@ interface IProps {
 
 const useStyles = makeStyles((theme) => ({
   paper: {
-    marginTop: theme.spacing(8),
+    marginTop: theme.spacing(2),
     display: 'flex',
     flexDirection: 'column',
     alignItems: 'center',
@@ -30,7 +30,7 @@ const useStyles = makeStyles((theme) => ({
     backgroundColor: theme.palette.secondary.main,
   },
   form: {
-    width: '100%', // Fix IE 11 issue.
+    width: '30%', // Fix IE 11 issue.
     marginTop: theme.spacing(1),
   },
   submit: {
