+++ src/features/WorkerScheduleForm/WorkerScheduleForm.tsx
@@ -4,13 +4,8 @@ import InputLabel from '@material-ui/core/InputLabel';
 import FormControl from '@material-ui/core/FormControl';
 import Select from '@material-ui/core/Select';
 import MenuItem from '@material-ui/core/MenuItem';
-import List from '@material-ui/core/List';
-import ListItem from '@material-ui/core/ListItem';
-import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction';
-import ListItemText from '@material-ui/core/ListItemText';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
-import TextField from '@material-ui/core/TextField';
 import Box from '@material-ui/core/Box';
 import Button from '@material-ui/core/Button';
 import Card from '@material-ui/core/Card';
@@ -27,26 +22,28 @@ import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
 import Checkbox from '@material-ui/core/Checkbox';
 
-import { IItemInfo, IWeekDays } from 'shared/types/models';
-import { workspacesInfo } from 'shared/workspaces';
+import { IWeekDays } from 'shared/types/models';
 
 const useStyles = makeStyles((theme) => ({
   formControl: {
     margin: theme.spacing(1),
     minWidth: 120,
+    maxWidth: 250,
+    'margin-left':  'auto',
+    'margin-right': 'auto',
   },
   progressBar: {
     margin: theme.spacing(1),
   },
   maingrid: {
     display: 'flex',
-    flexDirection: 'row',
+    flexDirection: 'column',
+    'justify-content': 'center',
   },
-  leftbox: {
-    width: '40%',
-  },
-  rightbox: {
-    width: '55%',
+  box: {
+    width: '50%',
+    'margin-left':  'auto',
+    'margin-right': 'auto',
   },
   timegrid: {
     display: 'flex',
@@ -55,15 +52,6 @@ const useStyles = makeStyles((theme) => ({
   card: {
     padding: theme.spacing(2, 2, 2),
   },
-  listdiv: {
-    backgroundColor: theme.palette.background.paper,
-  },
-  listitem: {
-    margin: 20,
-  },
-  title: {
-    margin: 10,
-  },
   subtitle: {},
   datepicker: {
     padding: 5,
@@ -71,10 +59,7 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 interface IProps {
-  workspace: string;
-
   usersList: string[];
-  itemsList: IItemInfo[];
 
   resText: string;
   isUserWaiting: boolean;
@@ -85,20 +70,14 @@ interface IProps {
   endDate: Date;
 
   handleUserChange: (event: React.ChangeEvent<{ value: unknown; }>) => void;
-  handleFieldChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
   handleWeekDaysChange: (name: string) => (event: React.ChangeEvent<HTMLInputElement>) => void;
   handleStartTimeChange: (d: Date | null) => void;
   handleEndTimeChange: (d: Date | null) => void;
-  handleItemsUpdateSubmit: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
   handleScheduleSubmit: (name: string) => (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
 }
 
 function mapToMenuItems(userList: string[]) {
   return userList.map(username => <MenuItem key={username} value={username}>{username}</MenuItem>);
-}
-
-function getItemTypeName(workspace: string, itemType: string) {
-  return workspacesInfo[workspace].types[itemType].description;
 }
 
 // TODO: import `freeText` from `ManagerFormContainer`.
@@ -115,10 +94,7 @@ const keyToLabel = {
 };
 
 const WorkerScheduleForm: React.FC<IProps> = ({
-  workspace,
-
   usersList,
-  itemsList,
 
   resText,
   isUserWaiting,
@@ -129,21 +105,15 @@ const WorkerScheduleForm: React.FC<IProps> = ({
   endDate,
 
   handleUserChange,
-  handleFieldChange,
   handleWeekDaysChange,
   handleStartTimeChange,
   handleEndTimeChange,
-  handleItemsUpdateSubmit,
   handleScheduleSubmit,
 }: IProps) => {
   const classes = useStyles();
 
   const timePickersDisabled = () => {
     return isUserWaiting && isTimeWaiting;
-  };
-
-  const itemsDisabled = () => {
-    return isUserWaiting;
   };
 
   const boxesIn = () => {
@@ -162,7 +132,7 @@ const WorkerScheduleForm: React.FC<IProps> = ({
 
   return (
     <MuiPickersUtilsProvider utils={DateFnsUtils}>
-      <div>
+      <Grid className={classes.maingrid}>
         <FormControl variant='outlined' className={classes.formControl}>
           <InputLabel ref={inputLabel} id='user-select-label'>
             {'User'}
@@ -186,58 +156,9 @@ const WorkerScheduleForm: React.FC<IProps> = ({
         >
           <CircularProgress className={classes.progressBar} />
         </Fade>
-        <Grid className={classes.maingrid}>
-          <Grow in={boxesIn()} /* TODO: MOVE TO ANOTHER TAB (new `feature`) */>
-            <Box component='span' m={1} className={classes.leftbox}>
-              <Card className={classes.card}>
-                <Grid item xs={10} md={11}>
-                  <Typography variant='h6' className={classes.title}>
-                    {'Available items'}
-                  </Typography>
-                  <div className={classes.listdiv}>
-                    <List>
-                      {itemsList.map(({ itemtype, count }) =>
-                        React.cloneElement(
-                          (
-                            <ListItem
-                              key={itemtype}
-                              className={classes.listitem}
-                            >
-                              <ListItemText primary={getItemTypeName(workspace, itemtype)} />
-                              <ListItemSecondaryAction>
-                                <TextField
-                                  id={itemtype}
-                                  label='Amount'
-                                  type='number'
-                                  variant='outlined'
-                                  value={count}
-                                  disabled={itemsDisabled()}
-                                  onChange={handleFieldChange}
-                                />
-                              </ListItemSecondaryAction>
-                            </ListItem>
-                          ) as React.ReactElement,
-                          {
-                            key: itemtype
-                          }
-                        )
-                      )}
-                    </List>
-                  </div>
-                  <Button
-                    variant='contained'
-                    color='primary'
-                    disabled={isUserWaiting}
-                    onClick={handleItemsUpdateSubmit}
-                  >
-                    {'Save'}
-                  </Button>
-                </Grid>
-              </Card>
-            </Box>
-          </Grow>
+        
           <Grow in={boxesIn()} {...(boxesIn() ? { timeout: 1000 } : {})}>
-            <Box component='span' m={1} className={classes.rightbox}>
+            <Box component='span' m={1} className={classes.box}>
               <Card className={classes.card}>
                 <Grid className={classes.timegrid}>
                   <div>
@@ -318,7 +239,6 @@ const WorkerScheduleForm: React.FC<IProps> = ({
             </Box>
           </Grow>
         </Grid>
-      </div>
     </MuiPickersUtilsProvider>
   );
 };
