+++ src/features/WorkspaceItemsForm/index.ts
@@ -1,0 +1,1 @@
+export { default as WorkspaceItemsForm } from './WorkspaceItemsFormContainer';
